Cleaned up some names and files
