--- messagestore_create_projector.go
@@ -1,5 +1,0 @@
-package gomessagestore
-
-func (ms *msgStore) CreateProjector() Projector {
-	return CreateProjector(ms.repo)
-}
