+++ repository/repo_read_stream_test.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	sqlmock "github.com/DATA-DOG/go-sqlmock"
+	. "github.com/blackhatbrigade/gomessagestore/repository"
 	"github.com/stretchr/testify/assert"
 	"golang.org/x/net/context"
 )
